5.0.2

+++ AbstractionLayer.IO/Contracts/IRenameQueue.cs
@@ -27,5 +27,5 @@ public interface IRenameQueue
     /// Executes the actual renaming.
     /// </summary>
     /// <returns>The number of files renamed.</returns>
-    ulong Commit();
+    int Commit();
 }
